requirements.txt
streamlit
pandas
langchain-groq
langchain-core
python-dotenv